Make flask testing optional
