+++ pytest_layab/flask.py
@@ -1,9 +1,21 @@
+import os
+import importlib
 import json
 from typing import Union
-import re
 
 from flask import Response
 from flask.testing import Client
+import pytest
+
+
+@pytest.fixture
+def app(service_module_name):
+    # Ensure that test configuration will be loaded
+    os.environ["SERVER_ENVIRONMENT"] = "test"
+    service_module = importlib.import_module(service_module_name)
+    service_module.application.testing = True
+    service_module.application.config["PROPAGATE_EXCEPTIONS"] = False
+    return service_module.application
 
 
 def assert_201(response: Response, expected_location: str) -> str:
@@ -22,48 +34,6 @@ def assert_201(response: Response, expected_location: str) -> str:
     return actual_location
 
 
-def assert_202_regex(response: Response, expected_location_regex: str) -> str:
-    """
-    Assert that status code is 202.
-    202 stands for Accepted, meaning that location header is expected as well.
-    Assert that location header is containing the expected location (hostname trimmed for tests)
-
-    :param response: response object from service to be asserted
-    :param expected_location_regex: Expected location starting from server root (eg: /xxx). Can be a regular exp.
-    :return Location from server root.
-    """
-    assert response.status_code == 202
-    actual_location = response.headers["location"].replace("http://localhost", "")
-    assert re.match(expected_location_regex, actual_location)
-    return actual_location
-
-
-def assert_204(response: Response):
-    """
-    Assert that status code is 204.
-    204 stands for No Content, meaning that body should be empty.
-    Assert that body is empty.
-    """
-    assert response.status_code == 204
-    assert response.data == b''
-
-
-def assert_303_regex(response: Response, expected_location_regex: str) -> str:
-    """
-    Assert that status code is 303.
-    303 stands for See Other, meaning that location header is expected as well.
-    Assert that location header is containing the expected location (hostname trimmed for tests)
-
-    :param response: response object from service to be asserted
-    :param expected_location_regex: Expected location starting from server root (eg: /xxx). Can be a regular exp.
-    :return Location from server root.
-    """
-    assert response.status_code == 303
-    actual_location = response.location.replace("http://localhost", "")
-    assert re.match(expected_location_regex, actual_location)
-    return actual_location
-
-
 def assert_file(response: Response, expected_file_path: str):
     """
     Assert that response is containing the bytes contained in expected file.
@@ -75,7 +45,9 @@ def assert_file(response: Response, expected_file_path: str):
         assert response.data == expected_file.read()
 
 
-def post_json(client: Client, url: str, json_body: Union[dict, list], **kwargs) -> Response:
+def post_json(
+    client: Client, url: str, json_body: Union[dict, list], **kwargs
+) -> Response:
     """
     Send a POST request to this URL.
 
@@ -114,7 +86,9 @@ def post_file(
         return client.post(url, data=data, **kwargs)
 
 
-def put_json(client: Client, url: str, json_body: Union[dict, list], **kwargs) -> Response:
+def put_json(
+    client: Client, url: str, json_body: Union[dict, list], **kwargs
+) -> Response:
     """
     Send a PUT request to this URL.
 
@@ -126,17 +100,3 @@ def put_json(client: Client, url: str, json_body: Union[dict, list], **kwargs) -
     return client.put(
         url, data=json.dumps(json_body), content_type="application/json", **kwargs
     )
-
-
-def assert_async(client: Client, response: Response) -> Response:
-    """
-    Ensure that received response is corresponding to an asynchronous request and wait for result.
-
-    :param client: Flask test client.
-    :param response: Received response.
-    :return: The final response containing the result.
-    """
-    status_url = assert_202_regex(response, ".*")
-    status_reply = client.get(status_url)
-    result_url = assert_303_regex(status_reply, ".*")
-    return client.get(result_url)
