--- tests/test_pytest_api_helper_failures.py
@@ -1,13 +1,0 @@
-import pytest
-from pytest_layab import service_module_name, before_service_init
-
-
-@pytest.fixture
-def test_module_name():
-    return "tests"
-
-
-# Ensure test will fail because fixture cannot be initialized
-@pytest.mark.xfail(reason="Expected failure because server.py file cannot be located.")
-def test_service_module_name_fixture(service_module_name):
-    pass
